Add files via upload
Files modified:
* www/background-mode.js
* src/android/ForegroundService.java

background-mode.js : The icon and color fields. were added to the settings JSON so they can be modified from the configuration on the javascript side.

ForegroundService.java : The color setting was added and also was the parsing from hex to integer, now the notification builder shows the color in it's background circle. The icon is also readed from the setting JSON and added to the same method that was used before, so it will be sought in the 'drawable' folder.